Don't clear BSS, it is already zeroed.
EFI doesn't seem to like a .bss section according to GNU-EFI.
Therefore it is put and expanded inside .data. It is already zeroed
there. It makes the EFI app bigger so maybe there is a better
way?

The reason clearBSS() made the application hang was because
kSystemTable that we set before clearBSS() was part of .bss which
meant it was erased on clearing.