bootloader: support booting an older kernel ABI
This allows the bootloader to boot a kernel which
uses an older kernel_args struct ABI.

The kernel_args, platform_args, and arch_args
structs have also introduced padding, so that
future expansion of these structs can avoid
further ABI changes.

A kernel args minor version field has also been
introduced, so when new fields are added to any
of these structs, the kernel can detect at
runtime whether the new fields will be
populated.

The reason this has been implemented in the boot
loader, rather than the kernel, is that an older
kernel won't be aware of the versioning
introduced with these changes.